% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.r
\name{BinaryCollider}
\alias{BinaryCollider}
\title{Explore bias due to sample ascertainment when the exposure and outcome are binary variables}
\description{
Explore bias due to sample ascertainment when the exposure and outcome are binary variables

Explore bias due to sample ascertainment when the exposure and outcome are binary variables
}
\details{
This function explores the range or OR generated by parameters in the following model:
P(S = 1 | A,Y) = b0 + ba\emph{A + by}Y + bay*AY
Where A is the exposure, Y is the outcome and S indicates whether an individual is present in the dataset
P(S = 1) is the proportion of the population that is present in the data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{param}}{Dataframe of parameter range list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-or_calc}{\code{BinaryCollider$or_calc()}}
\item \href{#method-ps_calc}{\code{BinaryCollider$ps_calc()}}
\item \href{#method-parameter_space}{\code{BinaryCollider$parameter_space()}}
\item \href{#method-scatter3d}{\code{BinaryCollider$scatter3d()}}
\item \href{#method-histogram}{\code{BinaryCollider$histogram()}}
\item \href{#method-clone}{\code{BinaryCollider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-or_calc"></a>}}
\subsection{Method \code{or_calc()}}{
Calculate the expected OR under collider bias when null hypothoses of OR = 1 is true:
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryCollider$or_calc(b0, ba, by, bay)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b0}}{Baseline probability of being selected}

\item{\code{ba}}{Effect of A on being selected}

\item{\code{by}}{Effect of Y on being selected}

\item{\code{bay}}{Effect of joint AY interaction on being selected}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Odds ratio
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ps_calc"></a>}}
\subsection{Method \code{ps_calc()}}{
Calculate the proportion of samples included in the model for a given set of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryCollider$ps_calc(b0, ba, by, bay, pA, pY, pAY)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b0}}{Baseline probability of being selected}

\item{\code{ba}}{Effect of A on being selected}

\item{\code{by}}{Effect of Y on being selected}

\item{\code{bay}}{Effect of joint AY interaction on being selected}

\item{\code{pA}}{P(A = 1) in the general population}

\item{\code{pY}}{P(Y = 1) in the general population}

\item{\code{pAY}}{P(A = 1, Y = 1) in the general population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
P(S = 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter_space"></a>}}
\subsection{Method \code{parameter_space()}}{
Specify a set of parameters for the structural equation, and calculate the set of odds ratios that would be obtained, assuming the odds ratio of A on Y in the total population is 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryCollider$parameter_space(
  target_or,
  pS,
  pA,
  pY,
  pAY,
  b0_range,
  ba_range,
  by_range,
  bay_range,
  granularity = 100,
  pS_tol = 0.0025
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_or}}{Target odds ratio. e.g. in an observational study this OR is observed, and the researcher seeks to find parameter ranges that could explain it}

\item{\code{pS}}{Proportion of the population present in the sample}

\item{\code{pA}}{P(A = 1) in the general population}

\item{\code{pY}}{P(Y = 1) in the general population}

\item{\code{pAY}}{P(A = 1, Y = 1) in the general population}

\item{\code{b0_range}}{Baseline probability of being selected. Provide a range of values to explore e.g. c(0,1)}

\item{\code{ba_range}}{Effect of A on being selected into the sample. Provide a range of values to explore e.g. c(-0.2, 0.2)}

\item{\code{by_range}}{Effect of Y on being selected into the sample. Provide a range of values to explore e.g. c(-0.2, 0.2)}

\item{\code{bay_range}}{Effect of AY interaction on being selected into the sample. Provide a range of values to explore e.g. c(-0.2, 0.2)}

\item{\code{granularity}}{Granularity of the search space. Default=100, going much higher can be computationally difficult}

\item{\code{pS_tol}}{Tolerance of pS value Default=0.0025}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of parameters that satisfy the target_or and target pS values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scatter3d"></a>}}
\subsection{Method \code{scatter3d()}}{
3D scatterplot of output from parameter_space function. See plot3D::scatter3D for info on parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryCollider$scatter3d(
  ticktype = "detailed",
  theta = 130,
  phi = 0,
  bty = "g",
  xlab = "ba",
  ylab = "by",
  zlab = "b0",
  clab = "OR",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticktype}}{Default="detailed"}

\item{\code{theta}}{Default=130}

\item{\code{phi}}{Default=0}

\item{\code{bty}}{Default="g"}

\item{\code{xlab}}{Default="ba"}

\item{\code{ylab}}{Default="by"}

\item{\code{zlab}}{Default="b0"}

\item{\code{clab}}{Default="OR"}

\item{\code{...}}{Further parameters to be passed to plot3D::scatter3D}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Scatterplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-histogram"></a>}}
\subsection{Method \code{histogram()}}{
Histogram of odds ratios across the range of parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryCollider$histogram(bins = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{How many bins to split histogram. Default=30}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinaryCollider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
