% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-structure.r
\name{plot_simulate_y_structure}
\alias{plot_simulate_y_structure}
\title{Plot Y structure bias across range of parameters}
\usage{
plot_simulate_y_structure(
  prop,
  b_xy_thresh,
  or_sz_range = c(1, 100),
  b_xy = 0,
  rsq_xs_range = c(0, 1),
  rsq_ys_range = c(0, 1),
  gran = 30
)
}
\arguments{
\item{prop}{Proportion of population included in sample}

\item{b_xy_thresh}{Target b_xy association - e.g. what value of b_xy are you suspicious could be due to ascertainment on a collider}

\item{or_sz_range}{Range of OR_sz values to enumerate over. Default = \code{c(1,100)}}

\item{b_xy}{Suspected true X-Y effect. Default = \code{0}}

\item{rsq_xs_range}{Range of rsq_xs values to enumerate over. Default = \code{c(0,1)}}

\item{rsq_ys_range}{Range of rsq_ys values to enumerate over. Default = \code{c(0,1)}}

\item{gran}{Granularity of ranges. Default = \code{30}}
}
\value{
ggplot of simulations
}
\description{
Takes a range of OR_sz values, and ranges of rsq_ys and rsq_xs values. Then, for a given proportion of the population being sampled, provides the expected X-Y association for each of the parameter combinations. Enumerates over all combinations so can be slow
}
