% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-structure.r
\name{plot_simulate_y_structure_optim}
\alias{plot_simulate_y_structure_optim}
\title{Plot the minimum Y-structure required to explain some association}
\usage{
plot_simulate_y_structure_optim(
  prop,
  b_xy_thresh,
  b_xy = 0,
  rsq_xs_range = c(0, 1),
  rsq_ys_range = c(0, 1),
  gran = 101,
  max_or_sz = 20
)
}
\arguments{
\item{prop}{Proportion of population included in sample}

\item{b_xy_thresh}{Target b_xy association - e.g. what value of b_xy are you suspicious could be due to ascertainment on a collider}

\item{b_xy}{Suspected true X-Y effect. Default = \code{0}}

\item{rsq_xs_range}{Range of rsq_xs values to enumerate over. Default = \code{c(0,1)}}

\item{rsq_ys_range}{Range of rsq_ys values to enumerate over. Default = \code{c(0,1)}}

\item{gran}{Granularity of ranges. Default = \code{101}}

\item{max_or_sz}{Maximum OR_sz to allow in optimisation.}
}
\value{
ggplot of simulations
}
\description{
For a given Y-structure, can an association of b_xy can be obtained through ascertainment on a collider? Use optimisation to determine the minimum OR_sz for a range of rsq_xs and rsq_ys values
}
