% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-structure.r
\name{simulate_y_structure}
\alias{simulate_y_structure}
\title{Y-structure collider}
\usage{
simulate_y_structure(
  n,
  prop,
  or_sz,
  b_xy,
  rsq_xs,
  rsq_ys,
  vx = 1,
  vy = 1,
  vs = 1,
  sig.level = 5e-08
)
}
\arguments{
\item{n}{Population sample size}

\item{prop}{Proportion of population in sample}

\item{or_sz}{Odds ratio for the effect of collider variable S on being included}

\item{b_xy}{Causal effect of X on Y}

\item{rsq_xs}{Variance explained by X on collider variable S}

\item{rsq_ys}{Variance explained by Y on collider variable S}

\item{vx}{Variance of X. Default = \code{1}.}

\item{vy}{Variance of Y. Default = \code{1}.}

\item{vs}{Variance of S. Default = \code{1}.}

\item{sig.level}{Alpha threshold for power calculation. Default = \code{5e-8}}
}
\value{
List of simulation results
}
\description{
X and Y are continuous, they influence continuous S which predicts binary inclusion variable Z
}
