% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.r
\name{VBB}
\alias{VBB}
\title{V-structure bias, binary exposure and binary outcome.}
\description{
V-structure bias, binary exposure and binary outcome.

V-structure bias, binary exposure and binary outcome.
}
\details{
Explore bias due to sample ascertainment when the exposure and outcome are binary variables. This class explores the range or OR generated by parameters in the following model:
\eqn{P(S = 1 | A,Y) = \beta_0 + \beta_a A + \beta_y Y + \beta_{ay}AY}
Where A is a binary exposure, Y is a binary outcome and S indicates whether an individual is present in the dataset
\eqn{P(S = 1)} is the proportion of the population that is present in the data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{param}}{Dataframe of parameter range list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-or_calc}{\code{VBB$or_calc()}}
\item \href{#method-ps_calc}{\code{VBB$ps_calc()}}
\item \href{#method-parameter_space}{\code{VBB$parameter_space()}}
\item \href{#method-scatter3d}{\code{VBB$scatter3d()}}
\item \href{#method-scatter}{\code{VBB$scatter()}}
\item \href{#method-histogram}{\code{VBB$histogram()}}
\item \href{#method-clone}{\code{VBB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-or_calc"></a>}}
\if{latex}{\out{\hypertarget{method-or_calc}{}}}
\subsection{Method \code{or_calc()}}{
Calculate the expected OR under collider bias when null hypothoses of OR = 1 is true:
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$or_calc(b0, ba, by, bay)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b0}}{Baseline probability of being selected}

\item{\code{ba}}{Effect of A on being selected}

\item{\code{by}}{Effect of Y on being selected}

\item{\code{bay}}{Effect of joint AY interaction on being selected}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Odds ratio
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ps_calc"></a>}}
\if{latex}{\out{\hypertarget{method-ps_calc}{}}}
\subsection{Method \code{ps_calc()}}{
Calculate the proportion of samples included in the model for a given set of parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$ps_calc(b0, ba, by, bay, pA, pY, pAY)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b0}}{Baseline probability of being selected}

\item{\code{ba}}{Effect of A on being selected}

\item{\code{by}}{Effect of Y on being selected}

\item{\code{bay}}{Effect of joint AY interaction on being selected}

\item{\code{pA}}{\eqn{P(A = 1)} in the general population}

\item{\code{pY}}{\eqn{P(Y = 1)} in the general population}

\item{\code{pAY}}{\eqn{P(A = 1, Y = 1)} in the general population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
P(S = 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter_space"></a>}}
\if{latex}{\out{\hypertarget{method-parameter_space}{}}}
\subsection{Method \code{parameter_space()}}{
Specify a set of parameters for the structural equation, and calculate the set of odds ratios that would be obtained, assuming the odds ratio of A on Y in the total population is 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$parameter_space(
  target_or,
  pS,
  pA,
  pY,
  pAY,
  b0_range,
  ba_range,
  by_range,
  bay_range,
  granularity = 100,
  pS_tol = 0.0025
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_or}}{Target odds ratio. e.g. in an observational study this OR is observed, and the researcher seeks to find parameter ranges that could explain it}

\item{\code{pS}}{Proportion of the population present in the sample}

\item{\code{pA}}{\eqn{P(A = 1)} in the general population}

\item{\code{pY}}{\eqn{P(Y = 1)} in the general population}

\item{\code{pAY}}{\eqn{P(A = 1, Y = 1)} in the general population}

\item{\code{b0_range}}{Baseline probability of being selected. Provide a range of values to explore e.g. \code{c(0,1)}}

\item{\code{ba_range}}{Effect of A on being selected into the sample. Provide a range of values to explore e.g. \code{c(-0.2, 0.2)}}

\item{\code{by_range}}{Effect of Y on being selected into the sample. Provide a range of values to explore e.g. \code{c(-0.2, 0.2)}}

\item{\code{bay_range}}{Effect of AY interaction on being selected into the sample. Provide a range of values to explore e.g. \code{c(-0.2, 0.2)}}

\item{\code{granularity}}{Granularity of the search space. Default=\code{100}, going much higher can be computationally difficult}

\item{\code{pS_tol}}{Tolerance of pS value Default=\code{0.0025}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame of parameters that satisfy the \code{target_or} and target \code{pS} values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scatter3d"></a>}}
\if{latex}{\out{\hypertarget{method-scatter3d}{}}}
\subsection{Method \code{scatter3d()}}{
3D scatterplot of output from \code{parameter_space} function. See \code{\link[plot3D:scatter3D]{plot3D::scatter3D}} for info on parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$scatter3d(
  ticktype = "detailed",
  theta = 130,
  phi = 0,
  bty = "g",
  xlab = "ba",
  ylab = "by",
  zlab = "b0",
  clab = "OR",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ticktype}}{Default=\code{"detailed"}}

\item{\code{theta}}{Default=\code{130}}

\item{\code{phi}}{Default=\code{0}}

\item{\code{bty}}{Default=\code{"g"}}

\item{\code{xlab}}{Default=\code{"ba"}}

\item{\code{ylab}}{Default=\code{"by"}}

\item{\code{zlab}}{Default=\code{"b0"}}

\item{\code{clab}}{Default=\code{"OR"}}

\item{\code{...}}{Further parameters to be passed to \code{\link[plot3D:scatter3D]{plot3D::scatter3D}}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Scatterplot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scatter"></a>}}
\if{latex}{\out{\hypertarget{method-scatter}{}}}
\subsection{Method \code{scatter()}}{
Simple scatterplot of output from parameter_space function. Plotted are the parameter values of \code{b0}, \code{ba} and \code{by} that can give rise to an OR \eqn{\ge} \code{target_or}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$scatter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-histogram"></a>}}
\if{latex}{\out{\hypertarget{method-histogram}{}}}
\subsection{Method \code{histogram()}}{
Histogram of odds ratios across the range of parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$histogram(bins = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{How many bins to split histogram. Default=\code{30}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VBB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
